chore: split integration tests
On-behalf-of: SAP <matthias.bruns@sap.com>
