--- bindings/go/transfer/integration/integration_test.go
@@ -1,1852 +1,0 @@
-package integration_test
-
-import (
-	"archive/tar"
-	"bytes"
-	"compress/gzip"
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"strings"
-	"testing"
-	"time"
-
-	awsconfig "github.com/aws/aws-sdk-go-v2/config"
-	awscreds "github.com/aws/aws-sdk-go-v2/credentials"
-	awss3 "github.com/aws/aws-sdk-go-v2/service/s3"
-	godigest "github.com/opencontainers/go-digest"
-	"github.com/opencontainers/image-spec/specs-go"
-	ocispecv1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/stretchr/testify/require"
-	"github.com/testcontainers/testcontainers-go"
-	"github.com/testcontainers/testcontainers-go/modules/minio"
-	"github.com/testcontainers/testcontainers-go/modules/registry"
-	"golang.org/x/crypto/bcrypt"
-	"oras.land/oras-go/v2"
-	"oras.land/oras-go/v2/content/memory"
-	"oras.land/oras-go/v2/registry/remote"
-	"oras.land/oras-go/v2/registry/remote/auth"
-	"oras.land/oras-go/v2/registry/remote/retry"
-
-	"ocm.software/open-component-model/bindings/go/blob/filesystem"
-	"ocm.software/open-component-model/bindings/go/blob/inmemory"
-	"ocm.software/open-component-model/bindings/go/credentials"
-	"ocm.software/open-component-model/bindings/go/ctf"
-	descriptor "ocm.software/open-component-model/bindings/go/descriptor/runtime"
-	descriptorv2 "ocm.software/open-component-model/bindings/go/descriptor/v2"
-	githubdigest "ocm.software/open-component-model/bindings/go/github/digest"
-	githubresource "ocm.software/open-component-model/bindings/go/github/repository/resource"
-	githubaccess "ocm.software/open-component-model/bindings/go/github/spec/access"
-	githubv1 "ocm.software/open-component-model/bindings/go/github/spec/access/v1"
-	"ocm.software/open-component-model/bindings/go/oci"
-	ocictf "ocm.software/open-component-model/bindings/go/oci/ctf"
-	"ocm.software/open-component-model/bindings/go/oci/repository/provider"
-	ociresource "ocm.software/open-component-model/bindings/go/oci/repository/resource"
-	urlresolver "ocm.software/open-component-model/bindings/go/oci/resolver/url"
-	ociaccessv1 "ocm.software/open-component-model/bindings/go/oci/spec/access/v1"
-	credidentity "ocm.software/open-component-model/bindings/go/oci/spec/identity/v1"
-	ctfrepospec "ocm.software/open-component-model/bindings/go/oci/spec/repository/v1/ctf"
-	ocirepospec "ocm.software/open-component-model/bindings/go/oci/spec/repository/v1/oci"
-	"ocm.software/open-component-model/bindings/go/repository"
-	"ocm.software/open-component-model/bindings/go/runtime"
-	s3repository "ocm.software/open-component-model/bindings/go/s3/repository"
-	s3access "ocm.software/open-component-model/bindings/go/s3/spec/access"
-	s3accessv1 "ocm.software/open-component-model/bindings/go/s3/spec/access/v1"
-	s3identityv1 "ocm.software/open-component-model/bindings/go/s3/spec/identity/v1"
-	s3v1alpha1 "ocm.software/open-component-model/bindings/go/s3/transformation/spec/v1alpha1"
-	"ocm.software/open-component-model/bindings/go/transfer"
-	transferv1alpha1 "ocm.software/open-component-model/bindings/go/transfer/v1alpha1/spec"
-	wgetrepository "ocm.software/open-component-model/bindings/go/wget/repository"
-	wgetaccess "ocm.software/open-component-model/bindings/go/wget/spec/access"
-	wgetaccessv1 "ocm.software/open-component-model/bindings/go/wget/spec/access/v1"
-	wgetv1alpha1 "ocm.software/open-component-model/bindings/go/wget/transformation/spec/v1alpha1"
-)
-
-const (
-	distributionRegistryImage = "registry:3.0.0"
-	testUsername              = "ocm"
-	testPassword              = "password"
-)
-
-// --- test helpers ---
-
-func startRegistry(t *testing.T) (address, user, password string) {
-	t.Helper()
-
-	htpasswd := generateHtpasswd(t, testUsername, testPassword)
-
-	ctx, cancel := context.WithTimeout(t.Context(), 60*time.Second)
-	defer cancel()
-
-	container, err := registry.Run(ctx, distributionRegistryImage,
-		registry.WithHtpasswd(htpasswd),
-		testcontainers.WithEnv(map[string]string{
-			"REGISTRY_VALIDATION_DISABLED": "true",
-			"REGISTRY_LOG_LEVEL":           "debug",
-		}),
-	)
-	require.NoError(t, err, "should start registry container")
-
-	t.Cleanup(func() {
-		if err := container.Terminate(context.Background()); err != nil {
-			t.Logf("failed to terminate registry container: %v", err)
-		}
-	})
-
-	addr, err := container.HostAddress(ctx)
-	require.NoError(t, err)
-
-	return addr, testUsername, testPassword
-}
-
-func generateHtpasswd(t *testing.T, username, password string) string {
-	t.Helper()
-	hash, err := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	require.NoError(t, err)
-	return fmt.Sprintf("%s:%s", username, string(hash))
-}
-
-func createAuthClient(address, username, password string) *auth.Client {
-	return &auth.Client{
-		Client:     retry.DefaultClient,
-		Credential: auth.StaticCredential(address, auth.Credential{Username: username, Password: password}),
-	}
-}
-
-// registryCreds holds credentials for a single OCI registry.
-type registryCreds struct {
-	address  string
-	username string
-	password string
-}
-
-// newCredResolver creates a credentials.StaticCredentialsResolver for one or more OCI registries.
-func newCredResolver(t *testing.T, registries ...registryCreds) *credentials.StaticCredentialsResolver {
-	t.Helper()
-	credMap := make(map[string]map[string]string)
-	for _, reg := range registries {
-		repo := &ocirepospec.Repository{
-			Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-			BaseUrl: fmt.Sprintf("http://%s", reg.address),
-		}
-		identity, err := credidentity.IdentityFromOCIRepository(repo)
-		require.NoError(t, err)
-		credMap[identity.String()] = map[string]string{
-			"username": reg.username,
-			"password": reg.password,
-		}
-	}
-	return credentials.NewStaticCredentialsResolver(credMap)
-}
-
-// createCTFRepository creates a CTF-backed OCI repository at the given path.
-func createCTFRepository(t *testing.T, path string, opts ...oci.RepositoryOption) repository.ComponentVersionRepository {
-	t.Helper()
-	fs, err := filesystem.NewFS(path, os.O_RDWR|os.O_CREATE)
-	require.NoError(t, err)
-	archive := ctf.NewFileSystemCTF(fs)
-	store := ocictf.NewFromCTF(archive)
-	allOpts := append([]oci.RepositoryOption{oci.WithResolver(store), oci.WithTempDir(t.TempDir())}, opts...)
-	repo, err := oci.NewRepository(allOpts...)
-	require.NoError(t, err)
-	return repo
-}
-
-// --- integration tests ---
-
-func Test_Integration_TransferLocalBlob_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start target OCI registry
-	registryAddr, user, password := startRegistry(t)
-
-	// 2. Create source CTF with a component version containing a local blob resource
-	componentName := "ocm.software/integration-test"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	resourceData := []byte("Hello, Integration Test!")
-	resourceBlob := inmemory.New(bytes.NewReader(resourceData))
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    componentName,
-					Version: componentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{Name: "test-resource", Version: "1.0.0"},
-					},
-					Type:     "plainText",
-					Relation: descriptor.LocalRelation,
-					Access: &descriptorv2.LocalBlob{
-						Type:      runtime.NewVersionedType(descriptorv2.LocalBlobAccessType, descriptorv2.LocalBlobAccessTypeVersion),
-						MediaType: "text/plain",
-					},
-				},
-			},
-		},
-	}
-
-	// Add resource to CTF — the returned resource has the updated access with localReference filled in
-	updatedResource, err := ctfRepo.AddLocalResource(t.Context(), componentName, componentVersion,
-		&desc.Component.Resources[0], resourceBlob)
-	r.NoError(err)
-	desc.Component.Resources[0] = *updatedResource
-
-	// Add component version to CTF
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// 3. Build the transfer graph
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err, "graph definition should build successfully")
-	r.NotNil(tgd)
-	r.NotEmpty(tgd.Transformations)
-
-	// 4. Build and execute the graph
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-
-	repoProvider := provider.NewComponentVersionRepositoryProvider(
-		provider.WithTempDir(t.TempDir()),
-	)
-	resourceRepo := ociresource.NewResourceRepository(nil)
-
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err, "graph should build and validate")
-
-	r.NoError(graph.Process(ctx), "graph execution should succeed")
-
-	// 5. Verify the component exists in the target registry
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Equal(componentVersion, gotDesc.Component.Version)
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("test-resource", gotDesc.Component.Resources[0].Name)
-}
-
-// addComponentWithResources creates a component descriptor with local blob resources,
-// adds them to the repo, and returns the descriptor.
-func addComponentWithResources(t *testing.T, repo repository.ComponentVersionRepository,
-	name, version string, resources map[string][]byte,
-) *descriptor.Descriptor {
-	t.Helper()
-	r := require.New(t)
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    name,
-					Version: version,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-		},
-	}
-
-	for resName, data := range resources {
-		res := descriptor.Resource{
-			ElementMeta: descriptor.ElementMeta{
-				ObjectMeta: descriptor.ObjectMeta{Name: resName, Version: "1.0.0"},
-			},
-			Type:     "plainText",
-			Relation: descriptor.LocalRelation,
-			Access: &descriptorv2.LocalBlob{
-				Type:      runtime.NewVersionedType(descriptorv2.LocalBlobAccessType, descriptorv2.LocalBlobAccessTypeVersion),
-				MediaType: "text/plain",
-			},
-		}
-
-		updatedResource, err := repo.AddLocalResource(t.Context(), name, version, &res, inmemory.New(bytes.NewReader(data)))
-		r.NoError(err)
-		desc.Component.Resources = append(desc.Component.Resources, *updatedResource)
-	}
-
-	r.NoError(repo.AddComponentVersion(t.Context(), desc))
-	return desc
-}
-
-func Test_Integration_TransferDescriptorOnly_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 2. Create source CTF with a component that has NO resources.
-	componentName := "ocm.software/descriptor-only"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    componentName,
-					Version: componentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// 3. Build the transfer graph.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// 4. Build and execute the graph.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 5. Verify the component exists in the target registry with correct metadata.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Equal(componentVersion, gotDesc.Component.Version)
-	r.Equal("test-provider", gotDesc.Component.Provider.Name)
-	r.Empty(gotDesc.Component.Resources, "descriptor-only component should have no resources")
-}
-
-func Test_Integration_TransferMultipleResources_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 2. Create source CTF with a component containing 3 resources.
-	componentName := "ocm.software/multi-resource"
-	componentVersion := "2.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	resources := map[string][]byte{
-		"resource-alpha": []byte("alpha content"),
-		"resource-beta":  []byte("beta content"),
-		"resource-gamma": []byte("gamma content"),
-	}
-	addComponentWithResources(t, ctfRepo, componentName, componentVersion, resources)
-
-	// 3. Build the transfer graph.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// 4. Build and execute the graph.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 5. Verify all 3 resources arrive in target.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Len(gotDesc.Component.Resources, 3, "all 3 resources should be transferred")
-
-	gotNames := make(map[string]bool)
-	for _, res := range gotDesc.Component.Resources {
-		gotNames[res.Name] = true
-	}
-	for name := range resources {
-		r.True(gotNames[name], "resource %q should exist in target", name)
-	}
-}
-
-func Test_Integration_TransferCTFToCTF(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Create source CTF with one component and one resource.
-	componentName := "ocm.software/ctf-to-ctf"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	addComponentWithResources(t, ctfRepo, componentName, componentVersion, map[string][]byte{
-		"my-resource": []byte("ctf-to-ctf data"),
-	})
-
-	// 2. Build the transfer graph with CTF target spec.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetCTFPath := t.TempDir()
-	targetSpec := &ctfrepospec.Repository{
-		Type:       runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath:   targetCTFPath,
-		AccessMode: "readwrite|create",
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// 3. Build and execute the graph (no credentials needed for CTF-to-CTF).
-	ctx := t.Context()
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, nil)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 4. Verify component arrives in target CTF.
-	targetRepo := createCTFRepository(t, targetCTFPath)
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target CTF")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Equal(componentVersion, gotDesc.Component.Version)
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("my-resource", gotDesc.Component.Resources[0].Name)
-}
-
-func Test_Integration_TransferMultipleComponents_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 2. Create source CTF with two different components.
-	component1Name := "ocm.software/multi-comp-alpha"
-	component1Version := "1.0.0"
-	component2Name := "ocm.software/multi-comp-beta"
-	component2Version := "2.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	addComponentWithResources(t, ctfRepo, component1Name, component1Version, map[string][]byte{
-		"alpha-res": []byte("alpha data"),
-	})
-	addComponentWithResources(t, ctfRepo, component2Name, component2Version, map[string][]byte{
-		"beta-res": []byte("beta data"),
-	})
-
-	// 3. Build the transfer graph with a single WithTransfer containing two Components.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{
-				{Component: component1Name, Version: component1Version},
-				{Component: component2Name, Version: component2Version},
-			},
-			Target:   targetSpec,
-			Resolver: transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// 4. Build and execute the graph.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 5. Verify both components arrive in target.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc1, err := targetRepo.GetComponentVersion(ctx, component1Name, component1Version)
-	r.NoError(err, "should find first component in target registry")
-	r.Equal(component1Name, gotDesc1.Component.Name)
-	r.Len(gotDesc1.Component.Resources, 1)
-	r.Equal("alpha-res", gotDesc1.Component.Resources[0].Name)
-
-	gotDesc2, err := targetRepo.GetComponentVersion(ctx, component2Name, component2Version)
-	r.NoError(err, "should find second component in target registry")
-	r.Equal(component2Name, gotDesc2.Component.Name)
-	r.Len(gotDesc2.Component.Resources, 1)
-	r.Equal("beta-res", gotDesc2.Component.Resources[0].Name)
-}
-
-func Test_Integration_TransferWithFromRepository(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 2. Create source CTF with a component and resource.
-	componentName := "ocm.software/from-repository"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	addComponentWithResources(t, ctfRepo, componentName, componentVersion, map[string][]byte{
-		"repo-resource": []byte("from-repository data"),
-	})
-
-	// 3. Build the transfer graph using NewRepositoryResolver instead of a full resolver.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// 4. Build and execute the graph.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 5. Verify the component exists in the target registry.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Equal(componentVersion, gotDesc.Component.Version)
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("repo-resource", gotDesc.Component.Resources[0].Name)
-}
-
-func Test_Integration_TransferRecursive_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 2. Create source CTF with a child component and a parent that references it.
-	childName := "ocm.software/recursive-child"
-	childVersion := "1.0.0"
-	parentName := "ocm.software/recursive-parent"
-	parentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	// Add child component (no resources).
-	childDesc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    childName,
-					Version: childVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), childDesc))
-
-	// Add parent component that references the child.
-	parentDesc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    parentName,
-					Version: parentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			References: []descriptor.Reference{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{
-							Name:    "child-ref",
-							Version: childVersion,
-						},
-					},
-					Component: childName,
-				},
-			},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), parentDesc))
-
-	// 3. Build the transfer graph with recursive enabled.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(),
-		&transferv1alpha1.Config{Recursive: transferv1alpha1.RecursiveInfinite},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: parentName, Version: parentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// 4. Build and execute the graph.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 5. Verify both parent and child arrive in the target.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotParent, err := targetRepo.GetComponentVersion(ctx, parentName, parentVersion)
-	r.NoError(err, "should find parent component in target registry")
-	r.Equal(parentName, gotParent.Component.Name)
-	r.Len(gotParent.Component.References, 1, "parent should have one reference")
-	r.Equal(childName, gotParent.Component.References[0].Component)
-
-	gotChild, err := targetRepo.GetComponentVersion(ctx, childName, childVersion)
-	r.NoError(err, "should find child component in target registry (recursive transfer)")
-	r.Equal(childName, gotChild.Component.Name)
-	r.Equal(childVersion, gotChild.Component.Version)
-}
-
-// pushTestOCIImage pushes a minimal OCI image (single layer) to the given registry.
-// Returns the full image reference (e.g., "localhost:5000/test/image:v1").
-func pushTestOCIImage(t *testing.T, registryAddr, user, password, repoPath, tag string) string {
-	t.Helper()
-
-	ref := fmt.Sprintf("%s/%s:%s", registryAddr, repoPath, tag)
-	// Return an http:// prefixed reference so the OCI client uses plain HTTP.
-	httpRef := fmt.Sprintf("http://%s", ref)
-
-	// Create a minimal OCI image: one layer + config + manifest.
-	layerContent := []byte("test layer content for integration test")
-	layerDesc := ocispecv1.Descriptor{
-		MediaType: ocispecv1.MediaTypeImageLayer,
-		Digest:    digestOf(layerContent),
-		Size:      int64(len(layerContent)),
-	}
-
-	configContent := []byte("{}")
-	configDesc := ocispecv1.Descriptor{
-		MediaType: ocispecv1.MediaTypeImageConfig,
-		Digest:    digestOf(configContent),
-		Size:      int64(len(configContent)),
-	}
-
-	manifest := ocispecv1.Manifest{
-		Versioned: specs.Versioned{SchemaVersion: 2},
-		MediaType: ocispecv1.MediaTypeImageManifest,
-		Config:    configDesc,
-		Layers:    []ocispecv1.Descriptor{layerDesc},
-	}
-
-	manifestContent, err := json.Marshal(manifest)
-	require.NoError(t, err)
-
-	manifestDesc := ocispecv1.Descriptor{
-		MediaType: ocispecv1.MediaTypeImageManifest,
-		Digest:    digestOf(manifestContent),
-		Size:      int64(len(manifestContent)),
-	}
-
-	// Push to an in-memory store, then copy to the remote registry.
-	store := memory.New()
-	ctx := t.Context()
-	require.NoError(t, store.Push(ctx, layerDesc, bytes.NewReader(layerContent)))
-	require.NoError(t, store.Push(ctx, configDesc, bytes.NewReader(configContent)))
-	require.NoError(t, store.Push(ctx, manifestDesc, bytes.NewReader(manifestContent)))
-	require.NoError(t, store.Tag(ctx, manifestDesc, tag))
-
-	repo, err := remote.NewRepository(ref)
-	require.NoError(t, err)
-	repo.PlainHTTP = true
-	repo.Client = &auth.Client{
-		Client:     retry.DefaultClient,
-		Credential: auth.StaticCredential(registryAddr, auth.Credential{Username: user, Password: password}),
-	}
-
-	_, err = oras.Copy(ctx, store, tag, repo, tag, oras.DefaultCopyOptions)
-	require.NoError(t, err, "should push test OCI image to source registry")
-
-	return httpRef
-}
-
-func digestOf(content []byte) godigest.Digest {
-	return godigest.FromBytes(content)
-}
-
-func Test_Integration_TransferOCIImageResource_CopyModeAllResources(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start source and target OCI registries.
-	sourceAddr, sourceUser, sourcePwd := startRegistry(t)
-	targetAddr, targetUser, targetPwd := startRegistry(t)
-
-	// 2. Push a test OCI image to the source registry.
-	imageRef := pushTestOCIImage(t, sourceAddr, sourceUser, sourcePwd, "test/image", "v1")
-
-	// 3. Create source CTF with a component that has an OCIImage resource
-	//    pointing to the image in the source registry.
-	componentName := "ocm.software/oci-resource-test"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    componentName,
-					Version: componentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{Name: "external-image", Version: "1.0.0"},
-					},
-					Type:     "ociImage",
-					Relation: descriptor.ExternalRelation,
-					Access: &ociaccessv1.OCIImage{
-						Type:           runtime.NewVersionedType(ociaccessv1.LegacyType, ociaccessv1.LegacyTypeVersion),
-						ImageReference: imageRef,
-					},
-				},
-			},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// 4. Build the transfer graph with CopyModeAllResources.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", targetAddr),
-	}
-
-	// Credential resolver that handles both source and target registries.
-	credResolver := newCredResolver(t,
-		registryCreds{sourceAddr, sourceUser, sourcePwd},
-		registryCreds{targetAddr, targetUser, targetPwd},
-	)
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(),
-		&transferv1alpha1.Config{CopyMode: transferv1alpha1.CopyModeAllResources},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// Verify that CopyModeAllResources generated OCI artifact transformations.
-	// With CopyModeLocalBlobResources, an OCIImage resource would be skipped entirely.
-	hasGetOCIArtifact := false
-	for _, tr := range tgd.Transformations {
-		if tr.Type.Name == "GetOCIArtifact" {
-			hasGetOCIArtifact = true
-			break
-		}
-	}
-	r.True(hasGetOCIArtifact, "CopyModeAllResources should generate GetOCIArtifact transformation for OCIImage resource")
-
-	// 5. Build and execute the graph.
-	ctx := t.Context()
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 6. Verify the component arrived in the target registry with the resource.
-	client := createAuthClient(targetAddr, targetUser, targetPwd)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(targetAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("external-image", gotDesc.Component.Resources[0].Name)
-
-	// Verify the resource was stored as a localBlob in the target registry.
-	gotAccess := gotDesc.Component.Resources[0].Access
-	r.NotNil(gotAccess, "resource access should not be nil")
-	r.Equal(descriptorv2.LocalBlobAccessType, gotAccess.GetType().Name,
-		"OCI image resource should be stored as localBlob in target after CopyModeAllResources transfer")
-
-	// Verify GlobalAccess is not set — transfer should produce a pure local blob without global access.
-	accessScheme := runtime.NewScheme(runtime.WithAllowUnknown())
-	descriptorv2.MustAddToScheme(accessScheme)
-	var typedLocalBlob descriptorv2.LocalBlob
-	r.NoError(accessScheme.Convert(gotAccess, &typedLocalBlob), "should convert access to LocalBlob")
-	r.Nil(typedLocalBlob.GlobalAccess, "localBlob should not have globalAccess after transfer")
-
-	// Verify the blob is actually present and readable in the target repository.
-	resourceIdentity := gotDesc.Component.Resources[0].ToIdentity()
-	localBlob, _, err := targetRepo.GetLocalResource(ctx, componentName, componentVersion, resourceIdentity)
-	r.NoError(err, "local blob should be retrievable from target repository")
-	reader, err := localBlob.ReadCloser()
-	r.NoError(err, "local blob should be readable")
-	defer func() { r.NoError(reader.Close()) }()
-	content, err := io.ReadAll(reader)
-	r.NoError(err)
-	r.NotEmpty(content, "local blob content should not be empty")
-}
-
-// Test_Integration_TransferOCIArtifact_OCIToOCI verifies the streaming TransferOCIArtifact path:
-// an OCIImage resource is transferred directly from one OCI registry to another without
-// intermediate tar materialisation. The resource access in the target descriptor must
-// be an OCI image reference, not a local blob.
-func Test_Integration_TransferOCIArtifact_OCIToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start source and target OCI registries.
-	sourceAddr, sourceUser, sourcePwd := startRegistry(t)
-	targetAddr, targetUser, targetPwd := startRegistry(t)
-
-	// 2. Push a test OCI image into the source registry.
-	imageRef := pushTestOCIImage(t, sourceAddr, sourceUser, sourcePwd, "test/image", "v1")
-
-	// 3. Build a component with an external OCIImage resource pointing at that image
-	//    and push it to a source OCI registry (not a CTF).
-	componentName := "ocm.software/streaming-transfer-test"
-	componentVersion := "1.0.0"
-
-	credResolver := newCredResolver(t,
-		registryCreds{sourceAddr, sourceUser, sourcePwd},
-		registryCreds{targetAddr, targetUser, targetPwd},
-	)
-
-	// Push the component descriptor to the source OCI registry.
-	sourceSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", sourceAddr),
-	}
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-
-	// Seed the source registry via the transfer builder itself.
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-	ctfSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    componentName,
-					Version: componentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{Name: "streamed-image", Version: "1.0.0"},
-					},
-					Type:     "ociImage",
-					Relation: descriptor.ExternalRelation,
-					Access: &ociaccessv1.OCIImage{
-						Type:           runtime.NewVersionedType(ociaccessv1.LegacyType, ociaccessv1.LegacyTypeVersion),
-						ImageReference: imageRef,
-					},
-				},
-			},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// CTF → source OCI (seed).
-	seedTGD, err := transfer.BuildGraphDefinition(t.Context(), nil,
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     sourceSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, ctfSpec),
-		},
-	)
-	r.NoError(err)
-	seedGraph, err := b.BuildAndCheck(seedTGD)
-	r.NoError(err)
-	r.NoError(seedGraph.Process(t.Context()))
-
-	// 4. Transfer OCI → OCI with UploadAsOciArtifact (streaming path).
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", targetAddr),
-	}
-
-	// Build a live repo client for the source OCI registry so it can be used as FromRepository.
-	sourceClient := createAuthClient(sourceAddr, sourceUser, sourcePwd)
-	sourceURLRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(sourceAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(sourceClient),
-	)
-	r.NoError(err)
-	sourceRepo, err := oci.NewRepository(oci.WithResolver(sourceURLRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(),
-		&transferv1alpha1.Config{
-			CopyMode:   transferv1alpha1.CopyModeAllResources,
-			UploadType: transferv1alpha1.UploadAsOciArtifact,
-		},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(sourceRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-
-	// Verify the graph contains a TransferOCIArtifact node, not GetOCIArtifact.
-	hasTransfer := false
-	for _, tr := range tgd.Transformations {
-		if tr.Type.Name == "TransferOCIArtifact" {
-			hasTransfer = true
-		}
-		r.NotEqual("GetOCIArtifact", tr.Type.Name,
-			"streaming path must not emit GetOCIArtifact")
-	}
-	r.True(hasTransfer, "UploadAsOciArtifact OCI→OCI must emit TransferOCIArtifact")
-
-	// Execute the graph.
-	streamGraph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(streamGraph.Process(t.Context()))
-
-	// 5. Verify the resource in the target has an OCI image reference (not a local blob).
-	client := createAuthClient(targetAddr, targetUser, targetPwd)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(targetAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(t.Context(), componentName, componentVersion)
-	r.NoError(err, "component must be present in target registry")
-	r.Len(gotDesc.Component.Resources, 1)
-
-	gotAccess := gotDesc.Component.Resources[0].Access
-	r.NotNil(gotAccess)
-
-	// Streaming transfer must produce an OCI image reference in the target, not a local blob.
-	r.Equal(ociaccessv1.LegacyType, gotAccess.GetType().Name,
-		"resource access must be OCIImage (not local blob) after streaming OCI-to-OCI transfer")
-
-	// Unmarshal the raw access to verify the image reference points at the target registry.
-	rawAccess, ok := gotAccess.(*runtime.Raw)
-	r.True(ok, "access must be a *runtime.Raw")
-	var ociAccess ociaccessv1.OCIImage
-	r.NoError(json.Unmarshal(rawAccess.Data, &ociAccess),
-		"resource access must unmarshal to OCIImage after streaming transfer")
-	r.Contains(ociAccess.ImageReference, targetAddr,
-		"image reference must point to the target registry")
-}
-
-// pushTestDockerManifest pushes a minimal Docker v2 manifest image to the given registry
-// and returns (imageRef http://addr/repo:tag, raw manifest bytes).
-func pushTestDockerManifest(t *testing.T, registryAddr, user, password, repoPath, tag string) (string, []byte) {
-	t.Helper()
-
-	const (
-		dockerManifestMediaType = "application/vnd.docker.distribution.manifest.v2+json"
-		dockerConfigMediaType   = "application/vnd.docker.container.image.v1+json"
-		dockerLayerMediaType    = "application/vnd.docker.image.rootfs.diff.tar.gzip"
-	)
-
-	ref := fmt.Sprintf("%s/%s:%s", registryAddr, repoPath, tag)
-
-	layerContent := []byte("docker layer content for integration test")
-	layerDesc := ocispecv1.Descriptor{
-		MediaType: dockerLayerMediaType,
-		Digest:    digestOf(layerContent),
-		Size:      int64(len(layerContent)),
-	}
-
-	configContent := []byte(`{"architecture":"amd64","os":"linux"}`)
-	configDesc := ocispecv1.Descriptor{
-		MediaType: dockerConfigMediaType,
-		Digest:    digestOf(configContent),
-		Size:      int64(len(configContent)),
-	}
-
-	// Docker manifest v2 uses a different mediaType from OCI.
-	type dockerManifest struct {
-		SchemaVersion int                    `json:"schemaVersion"`
-		MediaType     string                 `json:"mediaType"`
-		Config        ocispecv1.Descriptor   `json:"config"`
-		Layers        []ocispecv1.Descriptor `json:"layers"`
-	}
-	manifest := dockerManifest{
-		SchemaVersion: 2,
-		MediaType:     dockerManifestMediaType,
-		Config:        configDesc,
-		Layers:        []ocispecv1.Descriptor{layerDesc},
-	}
-	manifestContent, err := json.Marshal(manifest)
-	require.NoError(t, err)
-
-	manifestDesc := ocispecv1.Descriptor{
-		MediaType: dockerManifestMediaType,
-		Digest:    digestOf(manifestContent),
-		Size:      int64(len(manifestContent)),
-	}
-
-	store := memory.New()
-	ctx := t.Context()
-	require.NoError(t, store.Push(ctx, layerDesc, bytes.NewReader(layerContent)))
-	require.NoError(t, store.Push(ctx, configDesc, bytes.NewReader(configContent)))
-	require.NoError(t, store.Push(ctx, manifestDesc, bytes.NewReader(manifestContent)))
-	require.NoError(t, store.Tag(ctx, manifestDesc, tag))
-
-	repo, err := remote.NewRepository(ref)
-	require.NoError(t, err)
-	repo.PlainHTTP = true
-	repo.Client = &auth.Client{
-		Client:     retry.DefaultClient,
-		Credential: auth.StaticCredential(registryAddr, auth.Credential{Username: user, Password: password}),
-	}
-
-	_, err = oras.Copy(ctx, store, tag, repo, tag, oras.DefaultCopyOptions)
-	require.NoError(t, err, "should push Docker manifest to registry")
-
-	return fmt.Sprintf("http://%s", ref), manifestContent
-}
-
-// Test_Integration_TransferDockerManifestLocalBlob_CTFToOCI tests that a LocalBlob resource
-// with a Docker manifest media type is correctly transferred as an OCI artifact when using
-// UploadAsOciArtifact mode. This is a regression test for the bug where Docker manifests
-// were excluded from isOCICompliantManifest, causing them to be silently stored as local blobs.
-func Test_Integration_TransferDockerManifestLocalBlob_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	const dockerManifestMediaType = "application/vnd.docker.distribution.manifest.v2+json"
-
-	// 1. Start source and target registries.
-	sourceAddr, sourceUser, sourcePwd := startRegistry(t)
-	targetAddr, targetUser, targetPwd := startRegistry(t)
-
-	// 2. Push a Docker manifest image to the source registry.
-	imageRef, _ := pushTestDockerManifest(t, sourceAddr, sourceUser, sourcePwd, "test/docker-image", "v1")
-
-	// 3. Create source CTF with an OCIImage resource pointing at the Docker manifest in source registry.
-	componentName := "ocm.software/docker-manifest-test"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{Name: componentName, Version: componentVersion},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{Name: "docker-image", Version: "1.0.0"},
-					},
-					Type:     "ociImage",
-					Relation: descriptor.ExternalRelation,
-					Access: &ociaccessv1.OCIImage{
-						Type:           runtime.NewVersionedType(ociaccessv1.LegacyType, ociaccessv1.LegacyTypeVersion),
-						ImageReference: imageRef,
-					},
-				},
-			},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// 4. Transfer with CopyModeAllResources + UploadAsOciArtifact.
-	//    The Docker manifest OCIImage resource must be correctly transferred end-to-end.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", targetAddr),
-	}
-
-	credResolver := newCredResolver(t,
-		registryCreds{sourceAddr, sourceUser, sourcePwd},
-		registryCreds{targetAddr, targetUser, targetPwd},
-	)
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(),
-		&transferv1alpha1.Config{
-			CopyMode:   transferv1alpha1.CopyModeAllResources,
-			UploadType: transferv1alpha1.UploadAsOciArtifact,
-		},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// With CopyModeAllResources + UploadAsOciArtifact targeting an OCI registry, the graph
-	// takes the streaming path and emits TransferOCIArtifact (not GetOCIArtifact).
-	hasTransferOCIArtifact := false
-	for _, tr := range tgd.Transformations {
-		if tr.Type.Name == "TransferOCIArtifact" {
-			hasTransferOCIArtifact = true
-			break
-		}
-	}
-	r.True(hasTransferOCIArtifact, "Docker manifest OCIImage resource with CopyModeAllResources+UploadAsOciArtifact to OCI target should generate TransferOCIArtifact transformation")
-
-	// 5. Execute the transfer.
-	ctx := t.Context()
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := ociresource.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 6. Verify the component exists in the target and the resource has OCIImage access pointing to the target.
-	client := createAuthClient(targetAddr, targetUser, targetPwd)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(targetAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "transferred component should be in target registry")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Len(gotDesc.Component.Resources, 1)
-
-	// With UploadAsOciArtifact the Docker manifest image should be stored as an OCI artifact
-	// (OCIImage access) in the target, not as a local blob.
-	gotAccess := gotDesc.Component.Resources[0].Access
-	r.NotNil(gotAccess)
-	r.Equal(ociaccessv1.LegacyType, gotAccess.GetType().Name,
-		"Docker manifest resource should be stored as OCIImage access after CopyModeAllResources+UploadAsOciArtifact transfer")
-
-	var typedOCIAccess ociaccessv1.OCIImage
-	rawAccess, err := json.Marshal(gotAccess)
-	r.NoError(err)
-	r.NoError(json.Unmarshal(rawAccess, &typedOCIAccess))
-	r.Contains(typedOCIAccess.ImageReference, targetAddr,
-		"OCIImage access should reference the target registry after transfer")
-}
-
-// Test_Integration_TransferWgetResource_CopyModeAllResources verifies that a wget resource (content
-// behind an HTTP URL, not stored in the source CTF) is transferred by value: the content is
-// downloaded from the URL and embedded as a localBlob in the target OCI registry. wget resources are
-// external, so they are only copied under CopyModeAllResources.
-func Test_Integration_TransferWgetResource_CopyModeAllResources(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Serve the resource content over HTTP.
-	resourceData := []byte("Hello from wget integration test!")
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-		w.Header().Set("Content-Type", "text/plain")
-		_, _ = w.Write(resourceData)
-	}))
-	t.Cleanup(srv.Close)
-
-	// 2. Start the target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 3. Create the source CTF with a component whose resource points at the HTTP server.
-	componentName := "ocm.software/wget-resource-test"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	// The wget access lives in a separate binding the CTF/OCI repository scheme does not know,
-	// so store it pre-encoded as raw JSON (how descriptors carry access on disk). The transfer
-	// discovery decodes it via its own scheme, which registers the wget access type.
-	wgetAccess := &wgetaccessv1.Wget{
-		Type:      runtime.NewVersionedType(wgetaccess.WgetConsumerType, wgetaccessv1.Version),
-		URL:       srv.URL + "/artifact.txt",
-		MediaType: "text/plain",
-	}
-	rawWgetAccess := &runtime.Raw{}
-	r.NoError(runtime.NewScheme(runtime.WithAllowUnknown()).Convert(wgetAccess, rawWgetAccess))
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    componentName,
-					Version: componentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{Name: "wget-resource", Version: "1.0.0"},
-					},
-					Type:     "blob",
-					Relation: descriptor.ExternalRelation,
-					Access:   rawWgetAccess,
-				},
-			},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// 4. Build the transfer graph with CopyModeAllResources (external resources are skipped otherwise).
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(),
-		&transferv1alpha1.Config{CopyMode: transferv1alpha1.CopyModeAllResources},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// A wget resource should generate a DownloadWgetResource transformation.
-	hasDownloadWget := false
-	for _, tr := range tgd.Transformations {
-		if tr.Type.Name == wgetv1alpha1.DownloadWgetResourceType {
-			hasDownloadWget = true
-			break
-		}
-	}
-	r.True(hasDownloadWget, "wget resource should generate a DownloadWgetResource transformation")
-
-	// 5. Build and execute the graph.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	// A wget resource is downloaded via the wget resource repository. In the CLI this is a dynamic
-	// registry dispatching by access type; here the transfer only involves a wget resource, so the
-	// wget repository is the concrete repo to provide (mirrors the OCI tests passing the OCI repo).
-	resourceRepo := wgetrepository.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 6. Verify the component arrived in the target registry with the resource stored as a localBlob.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Equal(componentName, gotDesc.Component.Name)
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("wget-resource", gotDesc.Component.Resources[0].Name)
-
-	gotAccess := gotDesc.Component.Resources[0].Access
-	r.NotNil(gotAccess, "resource access should not be nil")
-	r.Equal(descriptorv2.LocalBlobAccessType, gotAccess.GetType().Name,
-		"wget resource should be stored as localBlob in target after transfer")
-
-	// Verify GlobalAccess is not set — transfer should produce a pure local blob without global access.
-	accessScheme := runtime.NewScheme(runtime.WithAllowUnknown())
-	descriptorv2.MustAddToScheme(accessScheme)
-	var typedLocalBlob descriptorv2.LocalBlob
-	r.NoError(accessScheme.Convert(gotAccess, &typedLocalBlob), "should convert access to LocalBlob")
-	r.Nil(typedLocalBlob.GlobalAccess, "localBlob should not have globalAccess after transfer")
-	r.Equal("text/plain", typedLocalBlob.MediaType, "localBlob should preserve the wget media type")
-
-	// The stored resource must be fully described: a digest over the downloaded content.
-	r.NotNil(gotDesc.Component.Resources[0].Digest, "transferred resource should carry a digest")
-	r.Equal(digestOf(resourceData).Encoded(), gotDesc.Component.Resources[0].Digest.Value,
-		"resource digest should be the sha256 of the served content")
-
-	// 7. Verify the blob is present and its content matches what the HTTP server served.
-	resourceIdentity := gotDesc.Component.Resources[0].ToIdentity()
-	localBlob, _, err := targetRepo.GetLocalResource(ctx, componentName, componentVersion, resourceIdentity)
-	r.NoError(err, "local blob should be retrievable from target repository")
-	reader, err := localBlob.ReadCloser()
-	r.NoError(err, "local blob should be readable")
-	defer func() { r.NoError(reader.Close()) }()
-	content, err := io.ReadAll(reader)
-	r.NoError(err)
-	r.Equal(resourceData, content, "transferred blob content should match the served content")
-}
-
-// The repository the mock GitHub API below serves. Nothing here reaches github.com.
-const (
-	ocmOwnerRepo = "open-component-model/open-component-model"
-	ocmCommit    = "f58349914e3c775747dc1ee9af1bc83db4652266"
-)
-
-// mockGitHubAPI serves the two endpoints a github download uses: the archive-link call, which
-// GitHub answers with a redirect, and the archive itself. The returned URL works verbatim as a
-// repoUrl because a non-github.com host makes the client treat it as GitHub Enterprise. The
-// archive path is matched by suffix, since go-github prefixes an Enterprise base URL with
-// /api/v3.
-func mockGitHubAPI(t *testing.T) (repoURL string, payload []byte) {
-	t.Helper()
-
-	var buf bytes.Buffer
-	gz := gzip.NewWriter(&buf)
-	tw := tar.NewWriter(gz)
-	const content = "hello world"
-	require.NoError(t, tw.WriteHeader(&tar.Header{
-		Name:     "open-component-model-open-component-model-" + ocmCommit + "/README",
-		Mode:     0o644,
-		Size:     int64(len(content)),
-		Typeflag: tar.TypeReg,
-	}))
-	_, err := tw.Write([]byte(content))
-	require.NoError(t, err)
-	require.NoError(t, tw.Close())
-	require.NoError(t, gz.Close())
-	payload = buf.Bytes()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		switch {
-		case strings.HasSuffix(req.URL.Path, "/repos/"+ocmOwnerRepo+"/tarball/"+ocmCommit):
-			http.Redirect(w, req, "http://"+req.Host+"/codeload", http.StatusFound)
-		case req.URL.Path == "/codeload":
-			_, _ = w.Write(payload)
-		default:
-			http.Error(w, "unexpected path "+req.URL.Path, http.StatusNotFound)
-		}
-	}))
-	t.Cleanup(server.Close)
-
-	return server.URL + "/" + ocmOwnerRepo, payload
-}
-
-// Transfers a gitHub resource pinned to a commit from a source CTF to an OCI registry, and
-// checks it lands in the target as a localBlob holding the repository archive.
-func Test_Integration_TransferGitHub_CTFToOCI(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-
-	// 1. Start the target OCI registry and the GitHub API the resource is fetched from.
-	registryAddr, user, password := startRegistry(t)
-	repoURL, archive := mockGitHubAPI(t)
-
-	// 2. Create a source CTF whose component has a gitHub resource pinned to a commit.
-	componentName := "ocm.software/github-integration-test"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	sourceScheme := runtime.NewScheme()
-	sourceScheme.MustRegisterScheme(oci.DefaultRepositoryScheme)
-	// The default repository scheme knows only OCI and localBlob, and this test hand-builds a
-	// typed *githubv1.GitHub access. The real authoring path does not need this: the CLI
-	// constructor converts external accesses to runtime.Raw before writing.
-	githubaccess.MustAddToScheme(sourceScheme)
-	ctfRepo := createCTFRepository(t, sourceCTFPath, oci.WithScheme(sourceScheme))
-
-	// A tgz of a repository tree is a directoryTree artifact; gitHub is the *access* type.
-	sourceResource := descriptor.Resource{
-		ElementMeta: descriptor.ElementMeta{
-			ObjectMeta: descriptor.ObjectMeta{Name: "repo-source", Version: "1.0.0"},
-		},
-		Type:     "directoryTree",
-		Relation: descriptor.ExternalRelation,
-		Access: &githubv1.GitHub{
-			Type:    runtime.NewVersionedType(githubv1.Type, githubv1.Version),
-			RepoURL: repoURL,
-			Commit:  ocmCommit,
-		},
-	}
-
-	// Pin the digest the way the constructor does. Without it the target would just compute a
-	// digest from whatever it stored, and the check below would never fail.
-	digested, err := githubdigest.NewDigestProcessor().ProcessResourceDigest(t.Context(), &sourceResource, nil)
-	r.NoError(err)
-	r.NotNil(digested.Digest, "digest processor must pin a digest")
-	pinnedDigest := digested.Digest.Value
-	r.Equal(digestOf(archive).Encoded(), pinnedDigest,
-		"the pinned digest must be the sha256 of the archive the API served")
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{Name: componentName, Version: componentVersion},
-			},
-			Provider:  descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{*digested},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(t.Context(), desc))
-
-	// 3. Build the transfer graph. CopyModeAllResources is required: the default
-	//    CopyModeLocalBlobResources skips external resources, and a gitHub access is external.
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(t.Context(),
-		&transferv1alpha1.Config{CopyMode: transferv1alpha1.CopyModeAllResources},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotEmpty(tgd.Transformations)
-
-	// 4. Build and execute the graph with the github resource repository.
-	ctx := t.Context()
-	credResolver := newCredResolver(t, registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	resourceRepo := githubresource.NewResourceRepository()
-
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 5. Verify the resource landed in the target as a localBlob.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err)
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("repo-source", gotDesc.Component.Resources[0].Name)
-	r.Equal(descriptorv2.LocalBlobAccessType,
-		gotDesc.Component.Resources[0].Access.GetType().Name,
-		"transferred github resource must be stored as a localBlob in the target")
-
-	// If the pinned digest changes, a signature over the source stops verifying against the
-	// transferred component version.
-	r.NotNil(gotDesc.Component.Resources[0].Digest, "transferred resource should carry a digest")
-	r.Equal(pinnedDigest, gotDesc.Component.Resources[0].Digest.Value,
-		"transferred resource must keep the digest pinned before the transfer")
-
-	// The stored bytes must be the archive the digest was taken over.
-	resourceIdentity := gotDesc.Component.Resources[0].ToIdentity()
-	localBlob, _, err := targetRepo.GetLocalResource(ctx, componentName, componentVersion, resourceIdentity)
-	r.NoError(err, "local blob should be retrievable from target repository")
-	reader, err := localBlob.ReadCloser()
-	r.NoError(err, "local blob should be readable")
-	defer func() { r.NoError(reader.Close()) }()
-	content, err := io.ReadAll(reader)
-	r.NoError(err)
-	r.Equal(archive, content, "stored blob must be the exact archive the GitHub API served")
-}
-
-// Test_Integration_TransferS3Resource_CopyModeAllResources verifies that an s3 resource (an
-// object in a bucket, not stored in the source CTF) is transferred by value: the object is
-// downloaded from the bucket and embedded as a localBlob in the target OCI registry. s3
-// resources are external, so they are only copied under CopyModeAllResources.
-func Test_Integration_TransferS3Resource_CopyModeAllResources(t *testing.T) {
-	t.Parallel()
-	r := require.New(t)
-	ctx := t.Context()
-
-	// 1. Serve the resource content from a MinIO bucket.
-	container, err := minio.Run(ctx, "minio/minio")
-	r.NoError(err)
-	t.Cleanup(func() { r.NoError(testcontainers.TerminateContainer(container)) })
-	hostPort, err := container.ConnectionString(ctx)
-	r.NoError(err)
-	endpoint := "http://" + hostPort
-
-	const bucket, objectKey = "transfer-bucket", "path/to/artifact.txt"
-	resourceData := []byte("Hello from s3 integration test!")
-	putS3Object(t, ctx, endpoint, container.Username, container.Password, bucket, objectKey, resourceData)
-
-	// 2. Start the target OCI registry.
-	registryAddr, user, password := startRegistry(t)
-
-	// 3. Create the source CTF with a component whose resource points at the bucket.
-	componentName := "ocm.software/s3-resource-test"
-	componentVersion := "1.0.0"
-	sourceCTFPath := t.TempDir()
-	ctfRepo := createCTFRepository(t, sourceCTFPath)
-
-	// The s3 access lives in a separate binding the CTF/OCI repository scheme does not know,
-	// so store it pre-encoded as raw JSON (how descriptors carry access on disk). The transfer
-	// discovery decodes it via its own scheme, which registers the s3 access type.
-	s3Access := &s3accessv1.S3Bucket{
-		Type:         s3access.V1VersionedType,
-		Region:       "us-east-1",
-		BucketName:   bucket,
-		ObjectKey:    objectKey,
-		MediaType:    "text/plain",
-		Endpoint:     endpoint,
-		UsePathStyle: true,
-	}
-	rawS3Access := &runtime.Raw{}
-	r.NoError(runtime.NewScheme(runtime.WithAllowUnknown()).Convert(s3Access, rawS3Access))
-
-	desc := &descriptor.Descriptor{
-		Meta: descriptor.Meta{Version: "v2"},
-		Component: descriptor.Component{
-			ComponentMeta: descriptor.ComponentMeta{
-				ObjectMeta: descriptor.ObjectMeta{
-					Name:    componentName,
-					Version: componentVersion,
-				},
-			},
-			Provider: descriptor.Provider{Name: "test-provider"},
-			Resources: []descriptor.Resource{
-				{
-					ElementMeta: descriptor.ElementMeta{
-						ObjectMeta: descriptor.ObjectMeta{Name: "s3-resource", Version: "1.0.0"},
-					},
-					Type:     "blob",
-					Relation: descriptor.ExternalRelation,
-					Access:   rawS3Access,
-				},
-			},
-		},
-	}
-	r.NoError(ctfRepo.AddComponentVersion(ctx, desc))
-
-	// 4. Build the transfer graph with CopyModeAllResources (external resources are skipped otherwise).
-	sourceSpec := &ctfrepospec.Repository{
-		Type:     runtime.Type{Name: ctfrepospec.Type, Version: ctfrepospec.Version},
-		FilePath: sourceCTFPath,
-	}
-	targetSpec := &ocirepospec.Repository{
-		Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-		BaseUrl: fmt.Sprintf("http://%s", registryAddr),
-	}
-
-	tgd, err := transfer.BuildGraphDefinition(ctx,
-		&transferv1alpha1.Config{CopyMode: transferv1alpha1.CopyModeAllResources},
-		transfer.Mapping{
-			Components: []transfer.ComponentID{{Component: componentName, Version: componentVersion}},
-			Target:     targetSpec,
-			Resolver:   transfer.NewRepositoryResolver(ctfRepo, sourceSpec),
-		},
-	)
-	r.NoError(err)
-	r.NotNil(tgd)
-
-	// An s3 resource should generate a DownloadS3Resource transformation.
-	hasDownloadS3 := false
-	for _, tr := range tgd.Transformations {
-		if tr.Type.Name == s3v1alpha1.DownloadS3ResourceType {
-			hasDownloadS3 = true
-			break
-		}
-	}
-	r.True(hasDownloadS3, "s3 resource should generate a DownloadS3Resource transformation")
-
-	// 5. Build and execute the graph.
-	credResolver := newS3CredResolver(t, endpoint, bucket, objectKey, container.Username, container.Password,
-		registryCreds{registryAddr, user, password})
-	repoProvider := provider.NewComponentVersionRepositoryProvider(provider.WithTempDir(t.TempDir()))
-	// An s3 resource is downloaded via the s3 resource repository. In the CLI this is a dynamic
-	// registry dispatching by access type; here the transfer only involves an s3 resource, so the
-	// s3 repository is the concrete repo to provide (mirrors the OCI tests passing the OCI repo).
-	resourceRepo := s3repository.NewResourceRepository(nil)
-	b := transfer.NewDefaultBuilder(repoProvider, resourceRepo, credResolver)
-	graph, err := b.BuildAndCheck(tgd)
-	r.NoError(err)
-	r.NoError(graph.Process(ctx))
-
-	// 6. Verify the component arrived in the target registry with the resource stored as a localBlob.
-	client := createAuthClient(registryAddr, user, password)
-	urlRes, err := urlresolver.New(
-		urlresolver.WithBaseURL(registryAddr),
-		urlresolver.WithPlainHTTP(true),
-		urlresolver.WithBaseClient(client),
-	)
-	r.NoError(err)
-	targetRepo, err := oci.NewRepository(oci.WithResolver(urlRes), oci.WithTempDir(t.TempDir()))
-	r.NoError(err)
-
-	gotDesc, err := targetRepo.GetComponentVersion(ctx, componentName, componentVersion)
-	r.NoError(err, "should find transferred component in target registry")
-	r.Len(gotDesc.Component.Resources, 1)
-	r.Equal("s3-resource", gotDesc.Component.Resources[0].Name)
-
-	gotAccess := gotDesc.Component.Resources[0].Access
-	r.NotNil(gotAccess, "resource access should not be nil")
-	r.Equal(descriptorv2.LocalBlobAccessType, gotAccess.GetType().Name,
-		"s3 resource should be stored as localBlob in target after transfer")
-
-	accessScheme := runtime.NewScheme(runtime.WithAllowUnknown())
-	descriptorv2.MustAddToScheme(accessScheme)
-	var typedLocalBlob descriptorv2.LocalBlob
-	r.NoError(accessScheme.Convert(gotAccess, &typedLocalBlob), "should convert access to LocalBlob")
-	r.Nil(typedLocalBlob.GlobalAccess, "localBlob should not have globalAccess after transfer")
-	r.Equal("text/plain", typedLocalBlob.MediaType, "localBlob should preserve the s3 media type")
-
-	// The stored resource must be fully described: a digest over the downloaded content.
-	r.NotNil(gotDesc.Component.Resources[0].Digest, "transferred resource should carry a digest")
-	r.Equal(digestOf(resourceData).Encoded(), gotDesc.Component.Resources[0].Digest.Value,
-		"resource digest should be the sha256 of the object content")
-
-	// 7. Verify the blob is present and its content matches the object in the bucket.
-	resourceIdentity := gotDesc.Component.Resources[0].ToIdentity()
-	localBlob, _, err := targetRepo.GetLocalResource(ctx, componentName, componentVersion, resourceIdentity)
-	r.NoError(err, "local blob should be retrievable from target repository")
-	reader, err := localBlob.ReadCloser()
-	r.NoError(err, "local blob should be readable")
-	defer func() { r.NoError(reader.Close()) }()
-	content, err := io.ReadAll(reader)
-	r.NoError(err)
-	r.Equal(resourceData, content, "transferred blob content should match the object content")
-}
-
-// putS3Object creates bucket in the store at endpoint and stores content under key.
-func putS3Object(t *testing.T, ctx context.Context, endpoint, user, pass, bucket, key string, content []byte) {
-	t.Helper()
-	cfg, err := awsconfig.LoadDefaultConfig(ctx,
-		awsconfig.WithRegion("us-east-1"),
-		awsconfig.WithCredentialsProvider(awscreds.NewStaticCredentialsProvider(user, pass, "")),
-	)
-	require.NoError(t, err)
-	client := awss3.NewFromConfig(cfg, func(o *awss3.Options) {
-		o.BaseEndpoint = &endpoint
-		o.UsePathStyle = true
-	})
-	_, err = client.CreateBucket(ctx, &awss3.CreateBucketInput{Bucket: &bucket})
-	require.NoError(t, err)
-	_, err = client.PutObject(ctx, &awss3.PutObjectInput{Bucket: &bucket, Key: &key, Body: bytes.NewReader(content)})
-	require.NoError(t, err)
-}
-
-// newS3CredResolver resolves the bucket credentials for the s3 consumer identity in addition to
-// the OCI registry credentials the transfer needs for its target.
-func newS3CredResolver(t *testing.T, endpoint, bucket, key, accessKeyID, secretAccessKey string, registries ...registryCreds) *credentials.StaticCredentialsResolver {
-	t.Helper()
-	credMap := make(map[string]map[string]string)
-	for _, reg := range registries {
-		repo := &ocirepospec.Repository{
-			Type:    runtime.Type{Name: ocirepospec.Type, Version: "v1"},
-			BaseUrl: fmt.Sprintf("http://%s", reg.address),
-		}
-		identity, err := credidentity.IdentityFromOCIRepository(repo)
-		require.NoError(t, err)
-		credMap[identity.String()] = map[string]string{
-			"username": reg.username,
-			"password": reg.password,
-		}
-	}
-	s3Identity, err := s3identityv1.IdentityFromObject(bucket, key, endpoint)
-	require.NoError(t, err)
-	credMap[s3Identity.String()] = map[string]string{
-		"accessKeyId":     accessKeyID,
-		"secretAccessKey": secretAccessKey,
-	}
-	return credentials.NewStaticCredentialsResolver(credMap)
-}
